dev-requirements.txt
-r requirements.txt
django-dotenv==1.3.0
ipython[notebook]
ipdb
django-pdb
django-debug-toolbar>=1.3.0